fix(screen logo): better logic in landing screen + better logo + logo on register
