+++ move_together_app/lib/views/register_screen.dart
@@ -2,9 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:move_together_app/Provider/auth_provider.dart';
+import 'package:move_together_app/Widgets/logo_area.dart';
 import 'package:move_together_app/core/services/auth_service.dart';
-
 import 'package:move_together_app/Widgets/button.dart';
+import 'package:move_together_app/Widgets/Input/cool_text_field.dart';
 
 class RegisterScreen extends StatefulWidget {
   const RegisterScreen({super.key});
@@ -29,7 +30,7 @@ class RegisterScreenState extends State<RegisterScreen> {
         setState(() {
           errorMessage = null;
         });
-        context.go('/login');
+        context.goNamed('login');
       }
     } catch (e) {
       if (mounted) {
@@ -43,27 +44,21 @@ class RegisterScreenState extends State<RegisterScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Register'),
-      ),
       body: Center(
         child: Padding(
           padding: const EdgeInsets.all(16.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              TextField(
+              const LogoArea(),
+              CoolTextField(
                 controller: _usernameController,
-                decoration: const InputDecoration(
-                  labelText: 'Username',
-                ),
+                hintText: 'Nom d\'utilisateur',
               ),
               const SizedBox(height: 16),
-              TextField(
+              CoolTextField(
                 controller: _passwordController,
-                decoration: const InputDecoration(
-                  labelText: 'Password',
-                ),
+                hintText: 'Mot de passe',
                 obscureText: true,
               ),
               const SizedBox(height: 16),
@@ -88,7 +83,7 @@ class RegisterScreenState extends State<RegisterScreen> {
                   ),
                   TextButton(
                     onPressed: () {
-                      context.goNamed('home');
+                      context.goNamed('landing');
                     },
                     child: const Text('Login'),
                   ),
